Reduce Restrict::restrict_to_inputs from n*n to n
It needs to compute the set of terms which only on inputs in S.

It did this via *separate* containment checks for each term.

But, of course, there is a simple graph traversal that does O(1) work
per term.

I'm taking this patch now b/c I need it. All credits to Evan.

It passes our ZK tests

Co-authored-by: Evan Laufer <evan.m.laufer@gmail.com>