Revert "Bugfixes: R1CS lowering (includes a soundness bugfix) (#124)"
This reverts commit 05b793d565c79e8a4b1650ec696fd537e97d04c7.

Right now Z#'s represents field comparisons and remainders
using 255-bit bit-vectors. Our field-blaster is not intended to be
correct or sound for bit-vectors that are this wide. Previously, a bug
in the field-blaster essentially allowed compilation and proving to
succeed anyway.

With the (reverted) bugfix, the field-blaster crashes.

Ultimately, we want Z# to handle these operations using
range-checks---not bit-vectors. So, I'm reverting the bugfix for now.